% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbh_class.R
\docType{class}
\name{GBH-class}
\alias{GBH-class}
\alias{initialize,GBH-method}
\alias{show,GBH-method}
\alias{GBH}
\alias{summary,GBH-method}
\alias{plot,GBH,ANY-method}
\alias{GBH,ANY}
\title{Class to facilitate performing the Group Benjamini-Hochberg procedure
  and interpreting its output.}
\usage{
\S4method{initialize}{GBH}(.Object, ...)

\S4method{show}{GBH}(object)

\S4method{summary}{GBH}(object)

\S4method{plot}{GBH,ANY}(x, ..., title = "GBH Adjustment")
}
\arguments{
\item{p.vals}{Object of class \code{"data.frame"}. Each row correpsonds to an
individual hypothesis. The first column stores the p-values before GBH
adjustment, while the second gives the GBH adjusted p-values. The
hypotheses are sorted in order of significance according to these GBH
adjusted p-values. The \code{group} column gives the group membership of
each hypothesis, and \code{adj.significnace} codes the significance of each
hypothesis, according to the GBH adjusted p-values.}

\item{pi0}{Object of class \code{"numeric"}. The proportion of null
hypotheses within each group. This is either known a priori or estimated
adaptively from the unadjusted p-values.}

\item{adaptive}{Object of class \code{"logical"}. An indicator of whether the
proportion \code{pi0} was estimated adaptively from the data or known a
priori.}

\item{alpha}{Object of class \code{"numeric"}. The level at which the FDR is
controlled, during the GBH procedure.}
}
\description{
Class to facilitate performing the Group Benjamini-Hochberg procedure
  and interpreting its output.
}
\section{Slots}{

\describe{
\item{\code{show}}{Prints the entire table of adjusted p-values and their associated
FDR adjusted significance levels, together with the estimated proportions
of null hypotheses, within each group.}

\item{\code{summary}}{Prints the most significant hypothesis, after adjusting for
multiple testing via GBH. Also supplies the estimated proportion of null
hypothesis within each group and a table of counts of adjusted significance
across groups.}

\item{\code{plot}}{Plots the p-values of the hypothesis, sorted according to GBH
adjusted significance, shape coded according to group membership, and color
coded according to pre and post GBH p-value adjustment.}
}}

\examples{
# These are the unadjusted p-values corresponding to the outcome of some
# multiple testing experiment. The first 500 hypotheses are null and the last
# 1500 are true alternatives.

unadjp <- c(runif(500, 0, 0.01), runif(1500, 0, 1))
names(unadjp) <- paste("Hyp: ", 1:2000)

# These are the unadjusted p-values corresponding to the outcome of some
# multiple testing experiment. The first 500 hypotheses are null and the last
# 1500 are true alternatives.
unadjp <- c(runif(500, 0, 0.01), runif(1500, 0, 1))
names(unadjp) <- paste("Hyp: ", 1:2000)

# Here there are two groups total we have randomly assigned hypotheses to these
# two groups.
group.index <- c(sample(1:2, 2000, replace = TRUE))

# Perform the GBH adjustment.
result <-  Adaptive.GBH(unadjp, group.index, method = "storey")

# A summary of the GBH adjustment
summary(result)
}
